requirements.txt
pymodbus
redis
python-can
modbus-tk
uvloop
twisted
pytz
gpiozero